requirements_doc.txt
sphinx
furo
numpydoc
sphinx-gallery
scipy
